Validate x_text_template with AGTextTemplate
